Remove BufferedProcessor from common reactive
