+++ common/reactive/src/main/java/io/helidon/common/reactive/MultiTappedProcessor.java
@@ -27,21 +27,24 @@ import java.util.function.Function;
  *
  * @param <R> Type of the processed items.
  */
-public class MultiTappedProcessor<R> extends BufferedProcessor<R, R> implements Multi<R> {
+public class MultiTappedProcessor<R> extends BaseProcessor<R, R> implements Multi<R> {
 
     private Optional<Function<R, R>> onNextFunction = Optional.empty();
     private Optional<Consumer<Throwable>> onErrorConsumer = Optional.empty();
     private Optional<Runnable> onCompleteRunnable = Optional.empty();
     private Optional<Consumer<Flow.Subscription>> onCancelConsumer = Optional.empty();
 
-    private MultiTappedProcessor() {
+    /**
+     * Create new processor with no functions to execute when signals are intercepted.
+     */
+    protected MultiTappedProcessor() {
     }
 
     /**
      * Create new processor with no functions to execute when signals are intercepted.
      *
      * @param <R> Type of the processed items.
-     * @return Brand new {@link MultiTappedProcessor}
+     * @return Brand new {@link io.helidon.common.reactive.MultiTappedProcessor}
      */
     public static <R> MultiTappedProcessor<R> create() {
         return new MultiTappedProcessor<>();
@@ -51,7 +54,7 @@ public class MultiTappedProcessor<R> extends BufferedProcessor<R, R> implements 
      * Set {@link java.util.function.Function} to be executed when onNext signal is intercepted.
      *
      * @param function Function to be invoked.
-     * @return This {@link MultiTappedProcessor}
+     * @return This {@link io.helidon.common.reactive.MultiTappedProcessor}
      */
     public MultiTappedProcessor<R> onNext(Function<R, R> function) {
         onNextFunction = Optional.ofNullable(function);
@@ -62,8 +65,8 @@ public class MultiTappedProcessor<R> extends BufferedProcessor<R, R> implements 
      * Set {@link java.util.function.Consumer} to be executed when onError signal is intercepted.
      *
      * @param consumer Consumer to be executed when onError signal is intercepted,
-     *                 argument is intercepted {@link java.lang.Throwable}.
-     * @return This {@link MultiTappedProcessor}
+     *                 argument is intercepted {@link Throwable}.
+     * @return This {@link io.helidon.common.reactive.MultiTappedProcessor}
      */
     public MultiTappedProcessor<R> onError(Consumer<Throwable> consumer) {
         onErrorConsumer = Optional.ofNullable(consumer);
@@ -71,10 +74,10 @@ public class MultiTappedProcessor<R> extends BufferedProcessor<R, R> implements 
     }
 
     /**
-     * Set {@link java.lang.Runnable} to be executed when onComplete signal is intercepted.
+     * Set {@link Runnable} to be executed when onComplete signal is intercepted.
      *
-     * @param runnable {@link java.lang.Runnable} to be executed.
-     * @return This {@link MultiTappedProcessor}
+     * @param runnable {@link Runnable} to be executed.
+     * @return This {@link io.helidon.common.reactive.MultiTappedProcessor}
      */
     public MultiTappedProcessor<R> onComplete(Runnable runnable) {
         onCompleteRunnable = Optional.ofNullable(runnable);
@@ -86,7 +89,7 @@ public class MultiTappedProcessor<R> extends BufferedProcessor<R, R> implements 
      *
      * @param consumer Consumer to be executed when onCancel signal is intercepted,
      *                 argument is intercepted {@link java.util.concurrent.Flow.Subscription}.
-     * @return This {@link MultiTappedProcessor}
+     * @return This {@link io.helidon.common.reactive.MultiTappedProcessor}
      */
     public MultiTappedProcessor<R> onCancel(Consumer<Flow.Subscription> consumer) {
         onCancelConsumer = Optional.ofNullable(consumer);
