+++ microprofile/reactive-streams/src/test/java/io/helidon/microprofile/reactive/ConsumableSubscriber.java
@@ -15,7 +15,7 @@
  *
  */
 
-package io.helidon.microrofile.reactive;
+package io.helidon.microprofile.reactive;
 
 import org.reactivestreams.Subscriber;
 import org.reactivestreams.Subscription;
