+++ microprofile/reactive-streams/src/test/java/io/helidon/microprofile/reactive/EngineTest.java
@@ -15,7 +15,7 @@
  *
  */
 
-package io.helidon.microrofile.reactive;
+package io.helidon.microprofile.reactive;
 
 import java.util.ArrayList;
 import java.util.Arrays;
