+++ microprofile/reactive-streams/src/test/java/io/helidon/microprofile/reactive/MultiLimitProcessorTest.java
@@ -15,7 +15,7 @@
  *
  */
 
-package io.helidon.microrofile.reactive;
+package io.helidon.microprofile.reactive;
 
 import java.util.Arrays;
 import java.util.List;
