+++ microprofile/reactive-streams/src/test/java/io/helidon/microprofile/reactive/MultiOnErrorResumeProcessorTest.java
@@ -15,7 +15,7 @@
  *
  */
 
-package io.helidon.microrofile.reactive;
+package io.helidon.microprofile.reactive;
 
 import java.util.Arrays;
 import java.util.Collections;
