+++ microprofile/reactive-streams/src/test/java/io/helidon/microprofile/reactive/TestThrowable.java
@@ -15,23 +15,7 @@
  *
  */
 
-package io.helidon.microrofile.reactive;
+package io.helidon.microprofile.reactive;
 
-import java.util.function.Function;
-
-import org.eclipse.microprofile.reactive.streams.operators.ReactiveStreams;
-import org.reactivestreams.Processor;
-
-public class MultiPeekProcessorTest extends AbstractProcessorTest {
-    @Override
-    protected Processor<Long, Long> getProcessor() {
-        return ReactiveStreams.<Long>builder().peek(i -> Function.identity()).buildRs();
-    }
-
-    @Override
-    protected Processor<Long, Long> getFailedProcessor(RuntimeException t) {
-        return ReactiveStreams.<Long>builder().peek(i -> {
-            throw t;
-        }).buildRs();
-    }
+public class TestThrowable extends Throwable {
 }
